renamed rollup -> rollup_node
