+++ pyproject.toml
@@ -49,7 +49,7 @@ deploy_erc20 = "scripts.etherlink:deploy_erc20"
 fund_etherlink_account = "scripts.etherlink:fund_account"
 deposit = "scripts.tezos:deposit"
 withdraw = "scripts.etherlink:withdraw"
-get_proof = "scripts.rollup:get_proof"
+get_proof = "scripts.rollup_node:get_proof"
 execute_outbox_message = "scripts.tezos:execute_outbox_message"
 parse_withdrawal_event = "scripts.etherlink:parse_withdrawal_event"
 build_tezos_contracts = "scripts.tezos:build_contracts"
@@ -57,6 +57,6 @@ build_etherlink_contracts = "scripts.etherlink:build_contracts"
 etherlink_tests = "scripts.etherlink:test_contracts"
 bootstrap = "scripts.bootstrap.bootstrap:rollout"
 bridge_token = "scripts.bridge_token:bridge_token"
-scan_outbox = "scripts.rollup:scan_outbox"
+scan_outbox = "scripts.rollup_node:scan_outbox"
 xtz_deposit = "scripts.tezos:xtz_deposit"
 xtz_withdraw = "scripts.etherlink:xtz_withdraw"
