+++ scripts/helpers/rollup_node/__init__.py
@@ -1,0 +1,15 @@
+from scripts.helpers.rollup_node.proof import (
+    Proof,
+    get_proof,
+    get_cemented_messages,
+    get_messages,
+)
+
+
+# Allowing reimporting from this module:
+__all__ = [
+    'Proof',
+    'get_proof',
+    'get_cemented_messages',
+    'get_messages',
+]
