--- scripts/rollup/__init__.py
@@ -1,9 +1,0 @@
-from scripts.rollup.get_proof import get_proof
-from scripts.rollup.scan_outbox import scan_outbox
-
-
-# Allowing reimporting from this module:
-__all__ = [
-    'get_proof',
-    'scan_outbox',
-]
