+++ scripts/rollup_node/__init__.py
@@ -1,0 +1,9 @@
+from scripts.rollup_node.get_proof import get_proof
+from scripts.rollup_node.scan_outbox import scan_outbox
+
+
+# Allowing reimporting from this module:
+__all__ = [
+    'get_proof',
+    'scan_outbox',
+]
